Fix build by making it use the (hardcoded) paths of dependencies.
The packages from our own apt repo are now installed as part of the
build process, and paths to these are simply hardcoded.

Had to upgrade allegro 5 and had to put some #define ASSERT(x) to make
it compile again. The reason for this is a bit unclear to me for now.