A few attempts to get V8 working correctly, for now it still compiles, possibly the issue is now in the API usage.
Need to compare our use of v8pp with the official examples, see if we
omit any function calls etc.

There was an issue at some point of me archiving the wrong .o files in
my custom v8.a, but now I no longer depend on making my own archive.

So I am more certain the issue is with the code right now.