Add SFML static lib to draw a real-time output.
The reason allegro5 cannot be used is that it is not thread-safe and
already used for the rendering itself.

SFML had to be patched to prevent a crash. XCloseDisplay() was a double
free because allegro5 probably already takes care of closing the
display

The new real time GUI is enabled with the existing --gui flag,
The old one is still available and renamed to --preview.