Support rendering to disk to compare performance.
However writing the files as frames seems to be slower for now.

Started troubleshooting why some messages get stuck in transit for a
long time > 10s in some cases when communicating with remote workers.

Debugging where CAF is wasting time has been hard and probably I will
replace the communication with the renderer <> workers with something
custom that is separate from CAF infrastructure.

CAF has been super convenient as an archtecture within a single process
for me, but with components running in different processes/machines I
find it hard to get a handle on it.