Draft implementation of manual render server <> clients.
Based on Beej's Guide to Network Programming

The goal is to see if there is a bottleneck with CAF.
This implementation of mine is already a few iterations now, and
contains a bunch of bugs that need to be solved.

This version for example gets stuck, and I learned from a few more
iterations of debugging that it is related to putting print statements
or not.