Create foundation for Motion Blur rendering, first version.
This implementation works on a limited set of movements but can be
extended in the future.

The mechanism is slightly complicated, but can be made more
understandable using future refactorings. Also I plan to document it in
the future since it's expected to be an important feature of this
rendering system.
The current code is ugly and extremely verbose, but it's the result of a
few busy evenings trying to get some proof of concept working.

Getting collision detection to work in tandem with this Motion blur
rendering system was a bit tricky to get right, but now that it works,
it also solves the issue you might have without considering Motion blur:
if objects move really fast, you might miss a collision if you're not
careful (depends on the method of course).

With this approach I decided to embrace V8 a bit more, instead of trying
to do as little as possible inside V8 and keep synchronizing objects
between C++ land and Javascript land.