Implement a way for objects to spawn child objects (in this.subobj) in real-time.
This is implemented by walking backwards through all three
instance/next/intermediate arrays so that we can insert new instances in
a linear way. Currently this does not support recursion yet.

For recursion we need to invoke init(), which can result in new objects
again. And call find_new_objects again, and again, until there is
nothing new to be created.

Example in motion4.js.