Draft technical documentation.
Make toggle for cursor in web UI. This allows for saving images without
the crosshair.

Bitmaps are now managed by the rendering engine itself.

There are now two bitmaps that the renderer uses to 100% correctly
calculate motion blur now.

Draw circle function returns bounding box of pixels touched, so that
renderer can render multiple passes of a circle (for motion blur) and
then once finished copy it to the final bmp. During the drawing process
the original background pixel needs to stay accessible, hence two the
need for two bitmaps now.

Fix client.cpp and make client_desktop working again.

Make the memory usage indicator display in megabytes instead of bytes in
the Turbo Vision UI.

Make the canvas resize with available width+height, and refactor
completely the viewpoint logic + client WASM. The canvas width and
height of the web UI is now leading. A future feature would be to
display a bounding box of the visible frame.

Fix my logger class to support flushing correctly, this became a problem
after integrating it in the tvision stuff.

The WASM Module is now constructed inside Vue space, as a workaround the
client.js had to be injected also dynamically after the Module object is
constructed.