Fix: motionblur_buffer - use ^2 and sqrt() to fix colors blending, and also fix small issue.
The workaround is great, but most likely this should be fixed in a different place, for now I will leave it as such.
Since I want to take a break of hours and hours of spending debugging color blending.

Fix: now use shape.scale again in drawing lines.

dunno why it was removed?

Some updates to javascripts while testing.

Some more sane defaults in main.cpp.