Draft a feature to connect lines and circles.
This required quite a few fixes, instance/next/intermediates should never point to the same instance.
For example, object with unique_id 1 is a different javascript object in all three of them.

Tried to fix a bunch of distance calculation issues, not everything is solved yet.

See triangles.js for an example of the work-in-progress feature of connecting lines and circles.

Left some debugging artifacts that are going to be helpful in the coming days further implementing this feature.