Comitting a half-baked attempt at implementing left[] / right[] feature.
Fixed build of emscripten and some minor other useful changes.

During this change, and debugging various issues that I don't fully
understand I came to the conclusion I will refactor a lot of stuff to
make it more comprehensible again.

For this reason a bunch of these changes will be undone anyway.