Separate v8_wrapper class because V8 can only be initialized once per process.
We want to create multiple instances of the Starcry class, e.g., in
integration tests, so the v8_wrapper / v8 context needs to be passed via
the constructor now as a dependency.

Starcry uses multi-threading, and to call into V8 requires an isolate to in the same thread.

Instead of creating a separate object for Isolates, decided to keep a single Isolate for now, and a member recreate_isolate_in_this_thread() instead.

This strange name is to make the purpose/intend very explicit.