Rewrite the positioning into something bug-free.
The new code should work correctly with different values for scale,
different sizes of the web UIs canvas, etc.

Temporarily removed the scale_ratio stuff, left it as comments in the
code, but not sure if it will be re-introduced.

The object labels are now a bit out-of-sync, but at least we can now
focus on fixing that knowing the positioning of the image works
correctly.

Also:

* Introduce debug flag to show some values used during rendering of chunks.

* Make number of chunks used for rendering configurable through web UI.

* Move debugging to container (port 18081 instead of 18080).