Remove workaround in test.js that were forcing the renderer to apply motion blur.
Some bugs have been fixed making this no longer needed.
It is still needed for the final scene in this script, where only the radius is expanding, and is not detected as motion. (code has been commented out in the past due to issues, but will soon be fixed hopefully)