Fix a bug if objects are spawned during script execution they could end up with the wrong parent.
The problem was that new objects are always inserted at the end of the
instances array, which is fine during initialization, but not if objects
that have already been instantiated are spawning new objects later.

Now we calculate correctly where to insert an instance, if an explicit
parent object is provided.

We also fixed a bug in another place where an assumption was made about
the ordering of objects ('next' vs. 'intermediate' instances),
introduced a simple unordered map which does the trick.

The input/script.js project seems to be doing good now.