Fix subtle issue of random 'black frames'.
This was a very subtle bug that happened in case the last attempt, during execution the steps half way need to be adjusted.
Normally this results in an early exit, and new attempt, but if it was the last attempt, this could cause issues with the stepper object.

We now freeze the stepper object if it's the last attempt, so no adjustments to it are made, and this issue is avoided