Copy generator class to native_generator, in preparation for implementing a native version.
The goal is to figure out how much performance is lost with our current V8 integration.

make builddev - will build a release build with DEVELOP=on
make debugdev - will build a debug build with DEVELOP=on

Let's see if we can develop it while keeping support for the old generator.