Restore the "props" field feature, which is a generic javascript object that should be able to hold anything.
Decided to keep it a v8::Local<v8::Object> in C++ land.
A lot of experimenting went before this, and I was not able to find a solution that could mimic the other fields, for example, this.props.foo = 'bar', would not trigger the property setter for 'props' in C++.
So we have to just set a flag when 'this.props' is accessed (triggers the getter), and then assume the user might have made some changes to it.