Fix an issue with "props" in case of restoring objects.
The properties were preserved+shared accross different states of objects (next/current/intermediate)
Likely because the shared pointer was being copied.
We have to use "recursively_copy_object", because just creating a new object and copying the fields with a simple "get+set" will not deep copy!

We do not always want this behavior, so instead of making a copy constructor that does this by default, we choose to make it an explicit function "reinitialize".