Refactor circle and ellipse drawing code.
Started by fixing some small bugs, re-learned how the code worked after many hours of debugging. Discovered the approach was not even correct.

A circle would be for example at position 0.5, 0.5 px (note it's a float). Then it would start at the center, and go using rel_x horizontally scanning pixels, and upwards using rel_y. At the same time it would mirror to the right, buttom and across. However, the values would be re-used, and incorrect, because the value on the left would not be 100% correct, we'd have a 0.5 rounding error. It generates slightly less beautiful movement in case objects are moving over the screen.

Anyway, decided to simplify+reimplement. Slightly slower code, but more correct, and now bug free, and easier to understand (IMO)