Add boost-ext::DI and started some code cleanup (attempt 2)
The culprit was that the generator instance had to be created as a
unique_ptr. Otherwise it would somehow try to re-use the instance,
causing all kinds of issues.
The second culprit, which I did found out about earlier was that the
shared_ptr references weren't playing along nicely with boost::di, so
getting rid of that fixed things as well.